fixed a function declaration

+++ src/cl_main.h
@@ -134,7 +134,7 @@ void				CLIENT_SetServerAddress( NETADDRESS_s Address );
 bool				CLIENT_GetAllowSendingOfUserInfo( void );
 void				CLIENT_SetAllowSendingOfUserInfo( bool bAllow );
 int					CLIENT_GetLatestServerGametic( void );
-void				CLIENT_SetLatestServerGametic( LONG latestServerGametic );
+void				CLIENT_SetLatestServerGametic( int latestServerGametic );
 unsigned int		CLIENT_GetEndFullUpdateTic( void );
 const FString		&CLIENT_GetPlayerAccountName( int player );
